requirements.txt
streamlit
python-dotenv
openai>=1.40.0
google-generativeai>=0.8.5
anthropic>=0.34.0
mistralai>=1.2.0
python-docx
pandas
